+++ Sources/VectorExtor/CGPath.Segment/CGPath.SegmentProtocol.swift
